# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "NID Mock Server API",
    "description": "API documentation for the NID Mock Server project. This API provides endpoints for managing National ID (NID) operations."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/id/{id}": {
      "get": {
        "tags": [
          "NID"
        ],
        "description": "Retrieve a specific National ID record by internal ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Internal ID of the NID record"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved NID record",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number",
                          "example": 1
                        },
                        "nid": {
                          "type": "string",
                          "example": "1234567890123"
                        },
                        "name": {
                          "type": "string",
                          "example": "John Doe"
                        },
                        "father_name": {
                          "type": "string",
                          "example": "Father Name"
                        },
                        "mother_name": {
                          "type": "string",
                          "example": "Mother Name"
                        },
                        "date_of_birth": {
                          "type": "string",
                          "example": "01/01/1990"
                        },
                        "address": {
                          "type": "string",
                          "example": "123 Main St, City, Country"
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number",
                          "example": 1
                        },
                        "nid": {
                          "type": "string",
                          "example": "1234567890123"
                        },
                        "name": {
                          "type": "string",
                          "example": "John Doe"
                        },
                        "father_name": {
                          "type": "string",
                          "example": "Father Name"
                        },
                        "mother_name": {
                          "type": "string",
                          "example": "Mother Name"
                        },
                        "date_of_birth": {
                          "type": "string",
                          "example": "01/01/1990"
                        },
                        "address": {
                          "type": "string",
                          "example": "123 Main St, City, Country"
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "404": {
            "description": "NID record not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "NID not found"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "NID not found"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Failed to fetch NID data"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Failed to fetch NID data"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/find": {
      "post": {
        "tags": [
          "NID"
        ],
        "description": "Find a National ID record by NID number and date of birth",
        "responses": {
          "200": {
            "description": "Successfully found NID record",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number",
                          "example": 1
                        },
                        "nid": {
                          "type": "string",
                          "example": "1234567890123"
                        },
                        "name": {
                          "type": "string",
                          "example": "John Doe"
                        },
                        "father_name": {
                          "type": "string",
                          "example": "Father Name"
                        },
                        "mother_name": {
                          "type": "string",
                          "example": "Mother Name"
                        },
                        "date_of_birth": {
                          "type": "string",
                          "example": "01/01/1990"
                        },
                        "address": {
                          "type": "string",
                          "example": "123 Main St, City, Country"
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number",
                          "example": 1
                        },
                        "nid": {
                          "type": "string",
                          "example": "1234567890123"
                        },
                        "name": {
                          "type": "string",
                          "example": "John Doe"
                        },
                        "father_name": {
                          "type": "string",
                          "example": "Father Name"
                        },
                        "mother_name": {
                          "type": "string",
                          "example": "Mother Name"
                        },
                        "date_of_birth": {
                          "type": "string",
                          "example": "01/01/1990"
                        },
                        "address": {
                          "type": "string",
                          "example": "123 Main St, City, Country"
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Missing required query parameters: nidnumber and dob"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Missing required query parameters: nidnumber and dob"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "404": {
            "description": "NID record not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "NID not found"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "NID not found"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Failed to fetch NID data"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Failed to fetch NID data"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "nidnumber",
                  "dob"
                ],
                "properties": {
                  "nidnumber": {
                    "type": "string",
                    "description": "National ID number",
                    "example": "1234567890123"
                  },
                  "dob": {
                    "type": "string",
                    "description": "Date of birth",
                    "example": "01/01/1990"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/find/{nidNumber}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "nidNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/search/name/{name}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/district/{district}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "district",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/division/{division}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "division",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/nid/": {
      "get": {
        "tags": [
          "NID"
        ],
        "description": "Retrieve all National ID records",
        "responses": {
          "200": {
            "description": "Successfully retrieved all NID records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "count": {
                      "type": "number",
                      "example": 150
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "example": 1
                          },
                          "nid": {
                            "type": "string",
                            "example": "1234567890123"
                          },
                          "name": {
                            "type": "string",
                            "example": "John Doe"
                          },
                          "father_name": {
                            "type": "string",
                            "example": "Father Name"
                          },
                          "mother_name": {
                            "type": "string",
                            "example": "Mother Name"
                          },
                          "date_of_birth": {
                            "type": "string",
                            "example": "01/01/1990"
                          },
                          "address": {
                            "type": "string",
                            "example": "123 Main St, City, Country"
                          }
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "count": {
                      "type": "number",
                      "example": 150
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "example": 1
                          },
                          "nid": {
                            "type": "string",
                            "example": "1234567890123"
                          },
                          "name": {
                            "type": "string",
                            "example": "John Doe"
                          },
                          "father_name": {
                            "type": "string",
                            "example": "Father Name"
                          },
                          "mother_name": {
                            "type": "string",
                            "example": "Mother Name"
                          },
                          "date_of_birth": {
                            "type": "string",
                            "example": "01/01/1990"
                          },
                          "address": {
                            "type": "string",
                            "example": "123 Main St, City, Country"
                          }
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Failed to fetch NID data"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Failed to fetch NID data"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/nid/find/{nidNumber}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "nidNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/nid/search/name/{name}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/nid/district/{district}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "district",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/nid/division/{division}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "division",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
...
}